+++ src/darwin.c
@@ -27,8 +27,9 @@ tr_error tr_fill_darwin(void *buf, size_t len)
 }
 
 #else
-tr_error tr_fill_darwin(void *buf [[maybe_unused]], size_t len [[maybe_unused]])
+tr_error tr_fill_darwin(void *buf, size_t len)
 {
+	(void)buf; (void)len;
 	return TR_ERR_NOT_SUPPORTED;
 }
 #endif