+++ src/urandom.c
@@ -38,8 +38,9 @@ tr_error tr_fill_urandom(void *buf, size_t len)
 }
 
 #else
-tr_error tr_fill_urandom(void *buf [[maybe_unused]], size_t len [[maybe_unused]])
+tr_error tr_fill_urandom(void *buf, size_t len)
 {
+	(void)buf; (void)len;
 	return TR_ERR_NOT_SUPPORTED;
 }
 #endif